maint: Upgrade vendored haml gem
The newer vendored version of rails includes deprecation warnings that
affect haml.  The upgrade gets rid of those for cleaner logs.

    DEPRECATION WARNING: Kernel#returning has been deprecated in favor of Object#tap

Haml now warns you when you start your server that you need to configure
sass to load if you want it to autocompile, but we ship a pregenerated
version of our CSS from sass so that you only need sass to develop, not
to run Dashboard.

Reviewed-by: