(#6828) Add a rake task for fast node deletion
Node deletion is really slow in the current version of dashboard because
it uses destroy instead of delete which creates a new activerecord model for
every record.

This has been fixed in the current code base (#6717), but not yet released.
This task is a stopgap until then and will never actually be merged in since
the regular node:del task will work once #6717 has been applied

Always remember to backup your data before running this.